# doc/api.spec.yaml
openapi: 3.0.3
info:
  title: Tournament API
  description: Table tennis tournament API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api/v1.0.0'
    description: Tournament api (local)
  - url: 'http://dev.tournament.api/api/v1.0.0'
    description: Tournament api (dev)
components:
  parameters:
...
    seriesId:
      name: seriesId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    offset:
      description: Pagination parameter for list operations
      name: offset
      in: query
      required: false
      schema:
        type: integer
    limit:
      description: Pagination parameter for list operations
      name: limit
      in: query
      required: false
      schema:
        type: integer
    sortBy:
      description: Order by field parameter for list operations
      name: sortBy
      in: query
      required: false
      schema:
        type: string
    sortOrder:
      description: Sort by ASC|DESC parameter for list operations
      name: sortOrder
      in: query
      required: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
  schemas:
    CommonResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
...
    SeriesRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum:
            - draft
            - in_progress
            - done
    SeriesResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - $ref: "#/components/schemas/SeriesRequest"
    TournamentRequest:
      type: object
      allOf:
        - $ref: "#/components/schemas/SeriesRequest"
      properties:
        seriesId:
          type: integer
          format: int64
        runDate:
          type: string
          format: date-time
    TournamentResponse:
      type: object
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - $ref: "#/components/schemas/TournamentRequest"
...
  examples:
...
    ExampleSeriesListResponse:
      description: List of all series
      value:
        - id: 1
          name: test series
          status: DRAFT
          created: 2021-09-28T08:00:00.000Z
          updated: 2021-09-28T08:00:00.000Z
        - id: 2
          name: test series 2
          created: 2021-09-28T08:00:00.000Z
          status: DRAFT
          updated: 2021-09-28T08:00:00.000Z
    ExampleSeriesResponse:
      description: Series created
      value:
        id: 1
        name: test series
        status: DRAFT
        created: 2021-09-28T08:00:00.000Z
        updated: 2021-09-28T08:00:00.000Z
    ExampleTournamentListResponse:
      description: List of all tournament
      value:
        - id: 1
          seriesId: 1
          runDate: 2021-09-28T08:00:00.000Z
          name: test tournament
          status: DRAFT
          created: 2021-09-28T08:00:00.000Z
          updated: 2021-09-28T08:00:00.000Z
        - id: 2
          seriesId: 1
          runDate: 2021-09-28T08:00:00.000Z
          name: test tournament 2
          status: DRAFT
          created: 2021-09-28T08:00:00.000Z
          updated: 2021-09-28T08:00:00.000Z
    ExampleTournamentResponse:
      description: Tournament created
      value:
        id: 1
        seriesId: 1
        runDate: 2021-09-28T08:00:00.000Z
        name: test tournament
        status: DRAFT
        created: 2021-09-28T08:00:00.000Z
        updated: 2021-09-28T08:00:00.000Z
...
paths:
...
  /series:
    get:
      tags:
        - series
      operationId: listSeries
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortOrder"
      description: List all series
      responses:
        '200':
          description: Successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeriesResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleSeriesListResponse"
    post:
      tags:
        - series
      operationId: createSeries
      description: Create new series
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/SeriesRequest"
      responses:
        '201':
          description: Created response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/SeriesResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleSeriesResponse"
  /series/{seriesId}:
    parameters:
      - $ref: "#/components/parameters/seriesId"
    get:
      tags:
        - series
      operationId: getSeries
      description: Get single series

      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/SeriesRequest"
    put:
      tags:
        - series
      operationId: updateSeries
      description: Update single series
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/SeriesRequest"
      responses:
        '200':
          description: Ok response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/SeriesResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleSeriesResponse"
    delete:
      tags:
        - series
      operationId: deleteSeries
      description: Remove series
      responses:
        '200':
          description: Ok response
        '404':
          description: Not Found
        '403':
          description: Forbidden

  ########################
  # tournament api
  ########################
  /series/{seriesId}/tournaments:
    get:
      tags:
        - series
      operationId: listTournaments
      parameters:
        - $ref: "#/components/parameters/seriesId"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortOrder"
      description: List all series
      responses:
        '200':
          description: Successful response
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TournamentResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleTournamentListResponse"
    post:
      tags:
        - series
      operationId: createTournament
      parameters:
        - $ref: "#/components/parameters/seriesId"
      description: Create new series
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/TournamentRequest"
      responses:
        '201':
          description: Created response
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/TournamentResponse"
              examples:
                response:
                  $ref: "#/components/examples/ExampleTournamentResponse"
...